+++ bzt/jmx.py
@@ -547,7 +547,7 @@ class JMX(object):
                                     testname=testname)
         udv_collection_prop = JMX._collection_prop("Arguments.arguments")
 
-        for var_name in sorted(udv_dict.keys(), key=lambda x: str(x)):
+        for var_name in sorted(udv_dict.keys(), key=str):
             udv_element_prop = JMX._element_prop(str(var_name), "Argument")
             udv_arg_name_prop = JMX._string_prop("Argument.name", var_name)
             udv_arg_value_prop = JMX._string_prop("Argument.value", udv_dict[var_name])
@@ -666,10 +666,15 @@ class JMX(object):
             parsed_url = parse.urlsplit(default_address)
             if parsed_url.scheme:
                 cfg.append(JMX._string_prop("HTTPSampler.protocol", parsed_url.scheme))
-            if parsed_url.hostname:
-                cfg.append(JMX._string_prop("HTTPSampler.domain", parsed_url.hostname))
-            if parsed_url.port:
-                cfg.append(JMX._string_prop("HTTPSampler.port", parsed_url.port))
+
+            if parsed_url.netloc:
+                netloc = parsed_url.netloc
+                if ':' in netloc:
+                    index = netloc.rfind(':')
+                    cfg.append(JMX._string_prop("HTTPSampler.port", netloc[index+1:]))
+                    netloc = netloc[:index]
+
+                cfg.append(JMX._string_prop("HTTPSampler.domain", netloc))
 
         if timeout:
             cfg.append(JMX._string_prop("HTTPSampler.connect_timeout", timeout))
@@ -942,7 +947,8 @@ class JMX(object):
         self.append(self.TEST_PLAN_SEL, dbg_tree)
         self.append(self.TEST_PLAN_SEL, etree.Element("hashTree"))
 
-    def _get_results_tree(self):
+    @staticmethod
+    def _get_results_tree():
         dbg_tree = etree.Element("ResultCollector",
                                  testname="View Results Tree",
                                  testclass="ResultCollector",